Work on gRPC implementation Add config support Add error handling Get owner from context (metadata)

+++ internal/maindb/maindb.go
@@ -2,13 +2,13 @@ package maindb
 
 import (
 	"context"
+	"github.com/Brialius/calendar/internal/domain/errors"
 	"github.com/Brialius/calendar/internal/domain/models"
 	_ "github.com/jackc/pgx/stdlib"
 	"github.com/jmoiron/sqlx"
 	"time"
 )
 
-// implements domain.interfaces.EventStorage
 type PgEventStorage struct {
 	db *sqlx.DB
 }
@@ -41,12 +41,22 @@ func (pges *PgEventStorage) SaveEvent(ctx context.Context, event *models.Event) 
 	return err
 }
 
-func (pges *PgEventStorage) GetEventById(ctx context.Context, id string) (*models.Event, error) {
-	// TODO
-	return nil, nil
+func (pges *PgEventStorage) GetEventByIdOwner(ctx context.Context, id, owner string) (*models.Event, error) {
+	query := `
+		SELECT * FROM events WHERE id=$1 AND owner=$2
+`
+	var events []*models.Event
+	err := pges.db.SelectContext(ctx, &events, query, id, owner)
+	if err != nil {
+		return nil, err
+	}
+	if len(events) == 1 {
+		return events[0], nil
+	}
+	return nil, errors.ErrNotFound
 }
 
-func (pges *PgEventStorage) GetEventsByOwnerStartDate(ctx context.Context, owner string, startTime time.Time) ([]*models.Event, error) {
+func (pges *PgEventStorage) GetEventsByOwnerStartDate(ctx context.Context, owner string, startTime *time.Time) ([]*models.Event, error) {
 	query := `
 		SELECT * FROM events WHERE owner=$1 AND start_time>=$2
 `
@@ -58,15 +68,38 @@ func (pges *PgEventStorage) GetEventsByOwnerStartDate(ctx context.Context, owner
 	return events, nil
 }
 
-func (pges *PgEventStorage) DeleteEventById(ctx context.Context, id string) error {
+func (pges *PgEventStorage) GetEventsCountByOwnerStartDateEndDate(ctx context.Context, owner string, startTime, endTime *time.Time) (int, error) {
 	query := `
-		DELETE FROM events WHERE id=$1
+		SELECT count(*) FROM events WHERE owner=$1 AND start_time>=$2 AND end_time<=$3
+`
+	var eventsCount int
+	err := pges.db.SelectContext(ctx, &eventsCount, query, owner, startTime, endTime)
+	if err != nil {
+		return eventsCount, err
+	}
+	return eventsCount, nil
+}
+
+func (pges *PgEventStorage) DeleteEventByIdOwner(ctx context.Context, id, owner string) error {
+	query := `
+		DELETE FROM events WHERE id=$1 AND owner=$2
 	`
-	_, err := pges.db.ExecContext(ctx, query, id)
+	res, err := pges.db.ExecContext(ctx, query, id, owner)
+	if res != nil {
+		if c, _ := res.RowsAffected(); c == 0 {
+			return errors.ErrNotFound
+		}
+	}
 	return err
 }
 
-func (pges *PgEventStorage) UpdateEventById(ctx context.Context, id string, event *models.Event) error {
-	// TODO
+func (pges *PgEventStorage) UpdateEventByIdOwner(ctx context.Context, id string, event *models.Event) error {
+	query := `
+		UPDATE events SET title=$3, text=$4, start_time=$5, end_time=$6 WHERE id=$1 AND owner=$2
+`
+	_, err := pges.db.ExecContext(ctx, query, id, event.Owner, event.Title, event.Text, event.StartTime, event.EndTime)
+	if err != nil {
+		return err
+	}
 	return nil
 }
